extract cardDisplayName helper fn
